Splitting capture and interactions
