+++ ui/src/home/home_page.jsx
@@ -5,7 +5,7 @@ import * as Colors from 'material-ui/styles/colors';
 import * as PropTypes from '../prop_types.js';
 import ChallengeCard from './challenge_card.jsx';
 import NavigationAppBar from '../components/navigation_app_bar.jsx';
-import Audio from '../components/audio.jsx';
+import AudioRecorder from '../components/audio_recorder.jsx';
 
 
 
@@ -40,7 +40,7 @@ export default React.createClass({
             })}
           </CardText>
         </Card>
-        <Audio />
+        <AudioRecorder />
       </div>
     );
   }
